requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
email_validator==2.2.0
razorpay==1.4.2
python-dotenv==1.1.0
Werkzeug==3.1.3
Jinja2==3.1.6
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0
sqlalchemy==2.0.40
wtforms==3.2.1
